Moved the child component to seperate file
