+++ frontend/src/components/MapView/MapView.tsx
@@ -1,12 +1,10 @@
 import { useCallback, useContext, useEffect, useState } from "react";
 import { MapContainer } from "react-leaflet/MapContainer";
-import { Marker } from "react-leaflet/Marker";
-import { Popup } from "react-leaflet/Popup";
+
 import { TileLayer } from "react-leaflet/TileLayer";
 import "leaflet/dist/leaflet.css";
 import "./MapView.css";
-import { useMap, useMapEvents } from "react-leaflet/hooks";
-import L, { DivIcon, LatLng } from "leaflet";
+import L, { LatLng } from "leaflet";
 import icon from "leaflet/dist/images/marker-icon.png";
 import iconShadow from "leaflet/dist/images/marker-shadow.png";
 import MapOptions from "./MapOptions";
@@ -16,9 +14,7 @@ import { MapContext } from "../../contexts/MapContext";
 import { TabProps, TabsContext } from "../../contexts/TabsContext";
 import { FeatureCollection } from "geojson";
 import { Dataset } from "../DatasetsList/DatasetsList";
-import { createRoot } from "react-dom/client";
-import { flushSync } from "react-dom";
-import { MapPin } from "@phosphor-icons/react";
+import MapEventsHandler from "./MapEventsHandler";
 
 const DefaultIcon = L.icon({
   iconUrl: icon,
@@ -27,23 +23,6 @@ const DefaultIcon = L.icon({
 });
 
 L.Marker.prototype.options.icon = DefaultIcon;
-
-// Utility function to render a React component to HTML string
-const renderToHtml = (Component: React.FC) => {
-  const div = document.createElement("div");
-  const root = createRoot(div);
-  flushSync(() => {
-    root.render(<Component />);
-  });
-  return div.innerHTML;
-};
-
-const divIconMarker: DivIcon = L.divIcon({
-  html: renderToHtml(() => <MapPin size={36} color="#ff0000" weight="fill" />),
-  className: "", // Optional: add a custom class name
-  iconSize: [36, 36],
-  iconAnchor: [18, 36], // Adjust the anchor point as needed
-});
 
 interface MapViewProps {
   datasetId: string;
@@ -95,62 +74,6 @@ const MapView: React.FC<MapViewProps> = ({ datasetId }) => {
       setCurrentMapCache((prev) => ({ ...prev, mapInstance: map }));
     }
   }, [map, setCurrentMapCache]);
-
-  const MapEventsHandler = () => {
-    const map = useMap();
-    useMapEvents({
-      click: (event) => {
-        console.log(event);
-        setCurrentMapCache({
-          ...currentMapCache,
-          selectedCoordinates: event.latlng,
-        });
-      },
-      moveend: (event) => {
-        setCurrentMapCache({
-          ...currentMapCache,
-          mapCenter: event.target.getCenter(),
-          mapBounds: event.target.getBounds(),
-          zoom: event.target.getZoom(),
-        });
-      },
-    });
-    return (
-      <Marker
-        position={currentMapCache.selectedCoordinates}
-        icon={divIconMarker}
-      >
-        <Popup>
-          <span
-            // Get the current location of the user
-            onClick={() => {
-              map
-                .locate({ setView: true })
-                .on("locationfound", function (event) {
-                  setPosition(event.latlng);
-                  map.flyTo(event.latlng, map.getZoom(), {
-                    animate: true,
-                    duration: 50,
-                  });
-                })
-                // If access to the location was denied
-                .on("locationerror", function (event) {
-                  console.log(event);
-                  alert("Location access denied.");
-                });
-            }}
-          >
-            {currentMapCache.selectedCoordinates.lat.toFixed(4)},{" "}
-            {currentMapCache.selectedCoordinates.lng.toFixed(4)}
-          </span>
-        </Popup>
-      </Marker>
-    );
-  };
-
-  function setPosition(latlng: L.LatLng) {
-    setCurrentMapCache({ ...currentMapCache, selectedCoordinates: latlng });
-  }
 
   // Get the feature collections from pinned tabs
   const pinnedFeatureCollections = currentTabsCache.openedTabs
